third_party/WebKit/LayoutTests/webaudio/Oscillator/oscillator-late-start-expected.txt

Test the late call of start(0) of Oscillator.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The index of first non-zero value is not equal to -1.
PASS The first sample value is equal to 0.
PASS The rendered buffer contains non-zero values after the first sample.
PASS successfullyParsed is true

TEST COMPLETE
